backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
